Improve inference
The original concept of this API is to work "over", "under", or
"a la" a *single* newtype, potentially changing its type arguments
but not changing it wholesale. Back in the day, Haskell couldn't
express that idea, but now it can (for the most part). We can take
advantage of that to improve inference and typed hole error messages.

Fixes #15