Improve inference (#16)
The original concept of this API is to work "over", "under", or
"a la" a *single* newtype, potentially changing its type arguments
but not changing it wholesale. Back in the day, Haskell couldn't
express that idea, but now it can (for the most part). We can take
advantage of that to improve inference and typed hole error messages.

Fixes #15

Also remove support for older GHC:

`Coercible` seems to be really fragile under GHC versions before
8.4. 8.2 seems especially fussy: even when the package compiles,
use cases may fail mysteriously.